Ai powered enhanced ehr/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
opencv-python-headless
pillow
numpy
pydantic
python-multipart       